refactor(perl-ci-hygiene): split process helpers into submodule (#0000)
